Added automatic mode. Need for bugfix
Currently the automatic mode kind of works, but eventually
a push is send to the server but either the revision
number or the local state ('awaitingPush') isn't reflected
properly, resultiung in an Unexpected ACK error message printed,
and a revision mismatch. We need to investigate
whether the local client state is ever modified in between the push
and the ack that we are awaiting for.